move config address type to address library
