+++ Makefile.autosetup
@@ -92,7 +92,8 @@ ALLOBJS+=	$(NEOMUTTOBJS)
 ###############################################################################
 # libaddress
 LIBADDRESS=	libaddress.a
-LIBADDRESSOBJS=	address/address.o address/group.o address/idna.o
+LIBADDRESSOBJS=	address/address.o address/config_type.o address/group.o \
+		address/idna.o
 CLEANFILES+=	$(LIBADDRESS) $(LIBADDRESSOBJS)
 ALLOBJS+=	$(LIBADDRESSOBJS)
 
@@ -263,9 +264,9 @@ $(PWD)/compress:
 ###############################################################################
 # libconfig
 LIBCONFIG=	libconfig.a
-LIBCONFIGOBJS=	config/address.o config/bool.o config/charset.o \
-		config/dump.o config/enum.o config/helpers.o config/long.o \
-		config/mbtable.o config/myvar.o config/number.o config/path.o \
+LIBCONFIGOBJS=	config/bool.o config/charset.o config/dump.o config/enum.o \
+		config/helpers.o config/long.o config/mbtable.o \
+		config/myvar.o config/number.o config/path.o \
 		config/quad.o config/regex.o config/set.o config/slist.o \
 		config/sort.o config/string.o config/subset.o
 CLEANFILES+=	$(LIBCONFIG) $(LIBCONFIGOBJS)
