--- config/address.h
@@ -1,31 +1,0 @@
-/**
- * @file
- * Type representing an email address
- *
- * @authors
- * Copyright (C) 2017-2018 Richard Russon <rich@flatcap.org>
- *
- * @copyright
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 2 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef MUTT_CONFIG_ADDRESS_H
-#define MUTT_CONFIG_ADDRESS_H
-
-struct Address;
-
-void            address_free(struct Address **ptr);
-struct Address *address_new (const char *addr);
-
-#endif /* MUTT_CONFIG_ADDRESS_H */
